backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
anthropic==0.36.0
chromadb==0.5.15
sentence-transformers==3.1.1
pypdf==4.3.1
python-docx==1.1.2
langchain-text-splitters==0.3.0
python-dotenv==1.0.1
pydantic==2.9.2
aiofiles==24.1.0
httpx==0.27.2
